+++ src/basic/view/view.py
@@ -21,14 +21,14 @@ class ParenthoodView(View):
         for c in self.childs.values(): c.draw()
 
 class Area(ABC):
-    def __init__(self, x: int, y: int, w: int, h: int):
+    def __init__(self, x: float, y: float, w: float, h: float):
         self.x = x
         self.y = y
         self.w = w
         self.h = h
 
 class CenteredArea(Area):
-    def to_center_pos(self, cx: int, cy: int):
+    def to_center_pos(self, cx: float, cy: float):
         self.x = cx - self.w / 2
         self.y = cy - self.h / 2
